requirements.txt
bm25s==0.2.7.post1
fastapi==0.115.11
llama-index==0.12.24
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.24.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.6.4.post1
loguru==0.7.3
numpy==1.26.4
openai==1.64.0
pandas==2.2.3
pyserini==0.44.0
pytrec_eval==0.5
requests==2.32.4
scikit-learn==1.6.1
scipy==1.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.56.1
trec-car-tools==2.6
ujson==5.10.0
